--- src/endpoint/index.js
@@ -1,3 +1,0 @@
-export { default as db } from './db';
-export { geocode, getLocations, reverseGeocode } from './geocode';
-export { default as handler, HttpException } from './handler';
