+++ src/pages/api/event/[id].js
@@ -1,4 +1,5 @@
-import { HttpException, db, getLocations, handler } from '../../../endpoint';
+import handler, { HttpException } from '../../handler';
+import getEvent from '../../../db/getEvent';
 
 /**
  * @param {import('http').IncomingMessage} req
@@ -8,60 +9,11 @@ const GET = async function GET(req, res) {
   const id = req.intParam('id');
   const limit = req.intParam('limit', null);
   const offset = req.intParam('offset', null);
-  const clauses = [];
-  const args = [id];
-
-  if (limit !== null) {
-    args.push(limit);
-    clauses.push(`LIMIT $${args.length}`);
-  }
-
-  if (offset !== null) {
-    args.push(offset);
-    clauses.push(`OFFSET $${args.length}`);
-  }
-
-  const query = `
-    SELECT
-      'Feature' as type,
-      jsonb_build_object(
-        'type', 'Point',
-        'coordinates', ARRAY[ST_X(location::geometry), ST_Y(location::geometry)]
-      ) as geometry,
-      jsonb_build_object(
-        'id', event.id,
-        'name', event.name,
-        'description', event.description,
-        'buskerId', busker.id,
-        'buskerName', busker.name,
-        'photos', coalesce(
-          array_agg(DISTINCT photo.url) FILTER (WHERE photo.url IS NOT NULL),
-          '{}'
-        ),
-        'tags', coalesce(
-          array_agg(DISTINCT tag.name) FILTER (WHERE tag.name IS NOT NULL),
-          '{}'
-        ),
-        'starts', event.starts,
-        'ends', event.ends
-      ) as properties
-  FROM event
-  LEFT JOIN event_photo ON event_photo.event_id = event.id
-  LEFT JOIN photo ON event_photo.photo_id = photo.id
-  LEFT JOIN event_tag ON event_tag.event_id = event.id
-  LEFT JOIN tag ON event_tag.tag_id = tag.id
-  JOIN busker ON busker.id = event.busker_id
-  WHERE event.id = $1
-  GROUP BY event.id, busker.id
-  ORDER BY starts ASC
-  ${clauses.join('\n')}
-  `;
-  const { rows } = await db.query(query, args);
-  if (rows.length === 0) {
+  const event = await getEvent(id, limit, offset);
+  if (event === undefined) {
     throw new HttpException(404, 'Event not found', id);
   }
-  await getLocations(rows);
-  res.status(200).json(rows[0]);
+  res.status(200).json(event);
 };
 
 export default handler({ GET });
