+++ src/pages/api/events.js
@@ -1,4 +1,12 @@
-import { HttpException, db, getLocations, handler } from '../../endpoint';
+import handler, { HttpException } from '../handler';
+import getEvents from '../../db/getEvents';
+
+const validFeatures = new Set([
+  'coords',
+  'location',
+  'photos',
+  'tags',
+]);
 
 /**
  * @param {import('http').IncomingMessage} req
@@ -14,29 +22,7 @@ const GET = async function GET(req, res) {
 
   const { features, sort } = req.query;
 
-  const clauses = [];
-  const wheres = [];
-  const args = [`SRID=4326;POINT(${lng} ${lat})`];
-
-  if (from !== null) {
-    args.push(from);
-    wheres.push(`starts >= $${args.length}`);
-  }
-
-  if (to !== null) {
-    args.push(to);
-    wheres.push(`ends <= $${args.length}`);
-  }
-
-  if (limit !== null) {
-    args.push(limit);
-    clauses.push(`LIMIT $${args.length}`);
-  }
-
-  if (offset !== null) {
-    args.push(offset);
-    clauses.push(`OFFSET $${args.length}`);
-  }
+  const parsedFeatures = features ? features.split(',') : [];
 
   let orderBy;
   if (sort === undefined || sort === 'distance') {
@@ -47,79 +33,26 @@ const GET = async function GET(req, res) {
     throw new HttpException(400, `unrecognized sort option ${sort}`, sort);
   }
 
-  let coordsQuery = '';
-  let photosQuery = '';
-  let photosJoin = '';
-  let tagsQuery = '';
-  let tagsJoin = '';
-  let getLocation = false;
+  const invalidFeatures = parsedFeatures
+    .filter(feature => !validFeatures.has(feature));
 
-  if (features !== undefined) {
-    for (const feature of features.split(',')) {
-      if (feature === 'coords') {
-        coordsQuery = `
-          'Feature' as type,
-          jsonb_build_object(
-            'type', 'Point',
-            'coordinates', ARRAY[ST_X(location::geometry), ST_Y(location::geometry)]
-          ) as geometry,
-        `;
-      } else if (feature === 'location') {
-        getLocation = true;
-      } else if (feature === 'photos') {
-        photosQuery = `
-          'photos', coalesce(array_agg(DISTINCT photo.url) FILTER (WHERE photo.url IS NOT NULL), '{}'),
-        `;
-        photosJoin = `
-          LEFT JOIN event_photo ON event_photo.event_id = event.id
-          LEFT JOIN photo ON event_photo.photo_id = photo.id
-        `;
-      } else if (feature === 'tags') {
-        tagsQuery = `
-          'tags', coalesce(array_agg(DISTINCT tag.name) FILTER (WHERE tag.name IS NOT NULL), '{}'),
-        `;
-        tagsJoin = `
-          LEFT JOIN event_tag ON event_tag.event_id = event.id
-          LEFT JOIN tag ON event_tag.tag_id = tag.id
-        `;
-      } else {
-        throw new HttpException(400, `unrecognized feature ${feature}`, feature);
-      }
-    }
+  if (invalidFeatures.length > 0) {
+    throw new HttpException(
+      400,
+      `unrecognized feature: ${invalidFeatures.join(', ')}`,
+      invalidFeatures,
+    );
   }
 
-  const query = `
-    SELECT
-      location <-> $1 AS distance,
-      ${coordsQuery}
-      jsonb_build_object(
-        'id', event.id,
-        'name', event.name,
-        'description', event.description,
-        'buskerId', busker.id,
-        'buskerName', busker.name,
-        ${photosQuery}
-        ${tagsQuery}
-        'starts', event.starts,
-        'ends', event.ends
-      ) as properties
-    FROM event
-    ${photosJoin}
-    ${tagsJoin}
-    JOIN busker ON busker.id = event.busker_id
-    ${wheres.length > 0 ? `WHERE ${wheres.join(' AND ')}` : ''}
-    GROUP BY event.id, busker.id
-    ORDER BY ${orderBy} ASC
-    ${clauses.join('\n    ')}
-  `;
-  const { rows } = await db.query(query, args);
-  if (getLocation) {
-    await getLocations(rows);
-  }
-  res.status(200).json({
-    type: 'FeatureCollection',
-    features: rows,
-  });
+  const events = await getEvents(
+    { lat, lng },
+    { from, to },
+    limit,
+    offset,
+    parsedFeatures,
+    orderBy,
+  );
+  res.status(200).json(events);
 };
 
 export default handler({ GET });
